PP-5758 Return 0 for refund amount available if not present
It is possible for the refund summary on a charge not to be
present when fetching a charge in process of refund. In that
case assume the amount available to refund is 0. This does mean
an extra call will be made to publicapi before the refund fails, but
has the advantage of keeping the logic contained in connector.